fix: Boot issues with ExitStrategy enum
Apparently with the way Spring constructs enums, we cannot provide a constructor to it as that data cannot be resolved at runtime. To fix this, ExitStrategy has been converted to a plain enum class, and a provider class ExitStrategyHandleProvider has been added to handle mapping the data from the configuration.